+++ packages/client/src/Ethereum.ts
@@ -44,10 +44,8 @@ const getOverrides = (chainName: string, provider: Provider, config: Pick<Strict
     const chainConfig = config.contracts.ethereumNetworks[chainName]
     if (chainConfig === undefined) { return {} }
     const overrides = chainConfig.overrides ?? {}
-    const gasPriceStrategy = chainConfig.highGasPriceStrategy
-        ? (estimatedGasPrice: BigNumber) => estimatedGasPrice.add('10000000000') 
-        : chainConfig.gasPriceStrategy
-    if (gasPriceStrategy !== undefined) {
+    if (chainConfig.highGasPriceStrategy) {
+        const gasPriceStrategy = (estimatedGasPrice: BigNumber) => estimatedGasPrice.add('10000000000') 
         return {
             ...overrides,
             gasPrice: provider.getGasPrice().then(gasPriceStrategy)